db test 14

+++ models/MYSQL_models/User_mdl.js
@@ -193,10 +193,13 @@ const User_model = {
             LEFT JOIN customer c ON u.user_id = c.customer_id_fk LEFT JOIN inventory_clerk ic ON u.user_id = ic.inventory_clerk_id_fk
             WHERE email = ? OR username = ?;`;
 
-            
-
-            console.log("CONNECTION DEAD",MySQL_DB.connection);
-            MySQL_DB.destroy();
+            if(MySQL_DB.connection)
+            {
+                console.log("CONNECTION DEAD",MySQL_DB.connection)
+                MySQL_DB.destroy()
+                .then(()=>{MySQL_DB.init();})
+                .catch((err)=>console.log(`Error in destroying DB`,err));
+            }
 
             db.connection.query(this.SQL, [user_email,user_username]) //customer_email,customer_username])
             .then(([rows,fields])=>{
